[INJIMOB-531] receive base64 encoded string to avoid data truncation (#46)
When we pass data from a react native app with <part1>\x00<part2> value, any data post the \x00 gets truncated (i.e., part2 is trimmed off). This happens most likely with the react native bridge. To avoid this issue of truncation, we are required to pass the data as base64 encoded string and secure-keystore module will perform decoding before processing for encryption.

Signed-off-by: KiruthikaJeyashankar <81218987+KiruthikaJeyashankar@users.noreply.github.com>
Co-authored-by: Harsh Vardhan <harsh59v@gmail.com>